exo-ingestion-gtfs/pyproject.toml
[project]
name = "exo_ingestion_gtfs"
version = "0.0.1"
description = "package responsable de l'ingestion des données brute gtfs produits par exo"
authors = ["Christian Bruneau <christian.bruneau777@gmail.com>"]

[project.dependencies]
requires-python = ">=3.10"

[tool.poetry]
name = "exo-ingestion-gtfs"
version = "0.1.0"
description = "exo gtfs ingestion"
authors = ["christian.bruneau"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
databricks-connect = "14.1.0"
fire = "0.5.0"
pandas = "1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
